+++ ShaderLibrary/Lighting.hlsl
@@ -21,25 +21,25 @@ SAMPLER(sampler_CellShading_Falloff);
 #define GRP_LIGHT_GET_TOTAL_FUNC DefaultGetLightTotal
 #endif
 
-float3 DefaultLightAttenuation(Surface surface, Light light)
+half3 DefaultLightAttenuation(Surface surface, Light light)
 {
-	float d = dot(surface.normal, light.direction);
+    half d = dot(surface.normal, light.direction);
 #if defined(_CELL_SHADING_ENABLED)
 	d = SAMPLE_TEXTURE2D(_CellShading_Falloff, sampler_CellShading_Falloff, float2(d, 0.5)).r;
 #endif
-    float a = lerp(1, saturate(d * light.attenuation), surface.shadowStrength);
+    half a = lerp(1, saturate(d * light.attenuation), surface.shadowStrength);
 	return light.color * a;
 }
 
-float3 DefaultGetLightTotal(Surface surface, BRDF brdf, GlobalIllumination globalIllumination)
+half3 DefaultGetLightTotal(Surface surface, BRDF brdf, GlobalIllumination globalIllumination)
 {
 	ShadowData shadowData = GetShadowData(surface);
 	
 	// Metallic, smoothness and fresnel
-	float3 indirect = IndirectBRDF(surface, brdf, globalIllumination.specular);
+    half3 indirect = IndirectBRDF(surface, brdf, globalIllumination.specular);
 	
 	// Direct lights
-	float3 result = indirect;
+    half3 result = indirect;
 	for (int i = 0; i < GetDirectionalLightCount(); i++)
 	{
 		Light light = GetDirectionalLight(i, surface, shadowData);
@@ -68,7 +68,7 @@ float3 DefaultGetLightTotal(Surface surface, BRDF brdf, GlobalIllumination globa
 	return result;
 }
 
-float GetPrimaryDirectionalShadow(float3 positionWS)
+half GetPrimaryDirectionalShadow(float3 positionWS)
 {
 	ShadowData global = GetShadowData(positionWS, -TransformWorldToView(positionWS).z, 0);
